+++ Cargo.toml
@@ -132,7 +132,7 @@ webpki-roots = "0.26"
 tokio-rustls = "0.25"
 native-tls = "0.2"
 tokio-native-tls = "0.3"
-tokio-tungstenite = "0.23.1"
+tokio-tungstenite = "0.23"
 
 [profile.release]
 opt-level = 3
