+++ examples/src/http/example-http-server.rs
@@ -12,13 +12,10 @@ use volo::{catch_panic, service::service_fn};
 use volo_http::{
     body::{Body, BodyConversion},
     context::{RequestPartsExt, ServerContext},
-    cookie::{self, CookieJar},
-    extension::Extension,
-    json::Json,
     request::ServerRequest,
     response::ServerResponse,
     server::{
-        extract::{Form, FromContext, MaybeInvalid, Query},
+        extract::{Form, FromContext, Json, MaybeInvalid, Query},
         layer::{FilterLayer, TimeoutLayer},
         middleware::{self, Next},
         panic_handler::{always_internal_error, fixed_payload},
@@ -27,6 +24,10 @@ use volo_http::{
         route::{get, get_service, post, Router},
         utils::ServeDir,
         IntoResponse, Redirect, Server,
+    },
+    utils::{
+        cookie::{self, CookieJar},
+        Extension,
     },
     Address,
 };
