+++ volo-http/src/server/route/router.rs
@@ -1,50 +1,29 @@
-//! Route module for routing path to [`Service`]s or handlers.
+//! [`Router`] implemententation for [`Server`].
 //!
-//! This module includes [`Router`], [`MethodRouter`] and [`Route`]. The call path is:
+//! [`Router`] is a [`Service`] that can route a path to a [`MethodRouter`] or fallback to another
+//! [`Route`].
 //!
-//! `Router` -> `MethodRouter` -> `Route`.
+//! See [`Router`] and [`Router::route`] for more details.
 //!
-//! [`Router`] is the main router for routing path (uri) to [`MethodRouter`]s. [`MethodRouter`] is
-//! a router for routing method (GET, POST, ...) to [`Route`]s. [`Route`] is a handler or service
-//! for handling the request.
+//! [`Server`]: crate::server::Server
 
-use std::{
-    collections::HashMap, convert::Infallible, error::Error, fmt, future::Future,
-    marker::PhantomData, str::FromStr,
-};
+use std::{collections::HashMap, convert::Infallible};
 
-use http::{Method, StatusCode, Uri};
+use http::status::StatusCode;
 use hyper::body::Incoming;
 use motore::{layer::Layer, service::Service, ServiceExt};
-use paste::paste;
 
-use super::{handler::Handler, IntoResponse};
-use crate::{context::ServerContext, request::ServerRequest, response::ServerResponse};
-
-/// The route service used for [`Router`].
-pub type Route<B = Incoming, E = Infallible> =
-    motore::service::BoxService<ServerContext, ServerRequest<B>, ServerResponse, E>;
-
-// The `matchit::Router` cannot be converted to `Iterator`, so using
-// `matchit::Router<MethodRouter>` is not convenient enough.
-//
-// To solve the problem, we refer to the implementation of `axum` and introduce a `RouteId` as a
-// bridge, the `matchit::Router` only handles some IDs and each ID corresponds to a `MethodRouter`.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-struct RouteId(u32);
-
-impl RouteId {
-    fn next() -> Self {
-        use std::sync::atomic::{AtomicU32, Ordering};
-        // `AtomicU64` isn't supported on all platforms
-        static ID: AtomicU32 = AtomicU32::new(0);
-        let id = ID.fetch_add(1, Ordering::Relaxed);
-        if id == u32::MAX {
-            panic!("Over `u32::MAX` routes created. If you need this, please file an issue.");
-        }
-        Self(id)
-    }
-}
+use super::{
+    method_router::MethodRouter,
+    utils::{Matcher, RouteId, StripPrefixLayer, NEST_CATCH_PARAM},
+    Fallback, Route,
+};
+use crate::{
+    context::ServerContext,
+    request::ServerRequest,
+    response::ServerResponse,
+    server::{handler::Handler, IntoResponse},
+};
 
 /// The router for routing path to [`Service`]s or handlers.
 #[must_use]
@@ -370,7 +349,7 @@ where
             is_default_fallback,
         } = other;
 
-        for (path, route_id) in matcher.matches.drain() {
+        for (path, route_id) in matcher.drain() {
             self.matcher
                 .insert_with_id(path, route_id)
                 .expect("Insert routing rule failed during merging router");
@@ -449,70 +428,10 @@ where
     }
 }
 
-#[derive(Default)]
-struct Matcher {
-    matches: HashMap<String, RouteId>,
-    router: matchit::Router<RouteId>,
-}
-
-impl Matcher {
-    fn insert<R>(&mut self, uri: R) -> Result<RouteId, MatcherError>
-    where
-        R: Into<String>,
-    {
-        let route_id = RouteId::next();
-        self.insert_with_id(uri, route_id)?;
-        Ok(route_id)
-    }
-
-    fn insert_with_id<R>(&mut self, uri: R, route_id: RouteId) -> Result<(), MatcherError>
-    where
-        R: Into<String>,
-    {
-        let uri = uri.into();
-        if self.matches.insert(uri.clone(), route_id).is_some() {
-            return Err(MatcherError::UriConflict(uri));
-        }
-        self.router
-            .insert(uri, route_id)
-            .map_err(MatcherError::RouterInsertError)?;
-        Ok(())
-    }
-
-    fn at<'a>(
-        &'a self,
-        path: &'a str,
-    ) -> Result<matchit::Match<'a, 'a, &'a RouteId>, MatcherError> {
-        self.router.at(path).map_err(MatcherError::RouterMatchError)
-    }
-}
-
-#[derive(Debug)]
-enum MatcherError {
-    UriConflict(String),
-    RouterInsertError(matchit::InsertError),
-    RouterMatchError(matchit::MatchError),
-}
-
-impl fmt::Display for MatcherError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::UriConflict(uri) => write!(f, "URI conflict: {uri}"),
-            Self::RouterInsertError(err) => write!(f, "router insert error: {err}"),
-            Self::RouterMatchError(err) => write!(f, "router match error: {err}"),
-        }
-    }
-}
-
-impl Error for MatcherError {}
-
 enum Endpoint<B = Incoming, E = Infallible> {
     MethodRouter(MethodRouter<B, E>),
     Service(Route<B, E>),
 }
-
-const NEST_CATCH_PARAM: &str = "{*__priv_nest_catch_param}";
-const NEST_CATCH_PARAM_NAME: &str = "__priv_nest_catch_param";
 
 impl<B, E> Service<ServerContext, ServerRequest<B>> for Endpoint<B, E>
 where
@@ -565,604 +484,22 @@ where
     }
 }
 
-/// A method router that handle the request and dispatch it by its method.
-///
-/// There is no need to create [`MethodRouter`] directly, you can use specific method for creating
-/// it. What's more, the method router allows chaining additional handlers or services.
-///
-/// # Examples
-///
-/// ```
-/// use std::convert::Infallible;
-///
-/// use volo::service::service_fn;
-/// use volo_http::{
-///     context::ServerContext,
-///     request::ServerRequest,
-///     server::route::{any, get, post_service, MethodRouter, Router},
-/// };
-///
-/// async fn index() -> &'static str {
-///     "Hello, World"
-/// }
-///
-/// async fn index_fn(
-///     cx: &mut ServerContext,
-///     req: ServerRequest,
-/// ) -> Result<&'static str, Infallible> {
-///     Ok("Hello, World")
-/// }
-///
-/// let _: MethodRouter = get(index);
-/// let _: MethodRouter = any(index);
-/// let _: MethodRouter = post_service(service_fn(index_fn));
-///
-/// let _: MethodRouter = get(index).post(index).options_service(service_fn(index_fn));
-///
-/// let app: Router = Router::new().route("/", get(index));
-/// let app: Router = Router::new().route("/", get(index).post(index).head(index));
-/// ```
-pub struct MethodRouter<B = Incoming, E = Infallible> {
-    options: MethodEndpoint<B, E>,
-    get: MethodEndpoint<B, E>,
-    post: MethodEndpoint<B, E>,
-    put: MethodEndpoint<B, E>,
-    delete: MethodEndpoint<B, E>,
-    head: MethodEndpoint<B, E>,
-    trace: MethodEndpoint<B, E>,
-    connect: MethodEndpoint<B, E>,
-    patch: MethodEndpoint<B, E>,
-    fallback: Fallback<B, E>,
-}
-
-impl<B, E> Service<ServerContext, ServerRequest<B>> for MethodRouter<B, E>
-where
-    B: Send,
-{
-    type Response = ServerResponse;
-    type Error = E;
-
-    async fn call(
-        &self,
-        cx: &mut ServerContext,
-        req: ServerRequest<B>,
-    ) -> Result<Self::Response, Self::Error> {
-        let handler = match *req.method() {
-            Method::OPTIONS => Some(&self.options),
-            Method::GET => Some(&self.get),
-            Method::POST => Some(&self.post),
-            Method::PUT => Some(&self.put),
-            Method::DELETE => Some(&self.delete),
-            Method::HEAD => Some(&self.head),
-            Method::TRACE => Some(&self.trace),
-            Method::CONNECT => Some(&self.connect),
-            Method::PATCH => Some(&self.patch),
-            _ => None,
-        };
-
-        match handler {
-            Some(MethodEndpoint::Route(route)) => route.call(cx, req).await,
-            _ => self.fallback.call(cx, req).await,
-        }
-    }
-}
-
-impl<B, E> Default for MethodRouter<B, E>
-where
-    B: Send + 'static,
-    E: 'static,
-{
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl<B, E> MethodRouter<B, E>
-where
-    B: Send + 'static,
-    E: 'static,
-{
-    fn new() -> Self {
-        Self {
-            options: MethodEndpoint::None,
-            get: MethodEndpoint::None,
-            post: MethodEndpoint::None,
-            put: MethodEndpoint::None,
-            delete: MethodEndpoint::None,
-            head: MethodEndpoint::None,
-            trace: MethodEndpoint::None,
-            connect: MethodEndpoint::None,
-            patch: MethodEndpoint::None,
-            fallback: Fallback::from_status_code(StatusCode::METHOD_NOT_ALLOWED),
-        }
-    }
-
-    /// Add a new inner layer to all routes in this method router.
-    ///
-    /// The layer's `Service` should be `Clone + Send + Sync + 'static`.
-    pub fn layer<L, B2, E2>(self, l: L) -> MethodRouter<B2, E2>
-    where
-        L: Layer<Route<B, E>> + Clone + Send + Sync + 'static,
-        L::Service: Service<ServerContext, ServerRequest<B2>, Error = E2> + Send + Sync + 'static,
-        <L::Service as Service<ServerContext, ServerRequest<B2>>>::Response: IntoResponse,
-        B2: 'static,
-    {
-        let Self {
-            options,
-            get,
-            post,
-            put,
-            delete,
-            head,
-            trace,
-            connect,
-            patch,
-            fallback,
-        } = self;
-
-        let layer_fn = move |route: Route<B, E>| {
-            Route::new(
-                l.clone()
-                    .layer(route)
-                    .map_response(IntoResponse::into_response),
-            )
-        };
-
-        let options = options.map(layer_fn.clone());
-        let get = get.map(layer_fn.clone());
-        let post = post.map(layer_fn.clone());
-        let put = put.map(layer_fn.clone());
-        let delete = delete.map(layer_fn.clone());
-        let head = head.map(layer_fn.clone());
-        let trace = trace.map(layer_fn.clone());
-        let connect = connect.map(layer_fn.clone());
-        let patch = patch.map(layer_fn.clone());
-
-        let fallback = fallback.map(layer_fn);
-
-        MethodRouter {
-            options,
-            get,
-            post,
-            put,
-            delete,
-            head,
-            trace,
-            connect,
-            patch,
-            fallback,
-        }
-    }
-}
-
-macro_rules! for_all_methods {
-    ($name:ident) => {
-        $name!(options, get, post, put, delete, head, trace, connect, patch);
-    };
-}
-
-macro_rules! impl_method_register_for_builder {
-    ($( $method:ident ),*) => {
-        $(
-        #[doc = concat!("Route `", stringify!($method) ,"` requests to the given handler.")]
-        pub fn $method<H, T>(mut self, handler: H) -> Self
-        where
-            for<'a> H: Handler<T, B, E> + Clone + Send + Sync + 'a,
-            B: Send,
-            T: 'static,
-        {
-            self.$method = MethodEndpoint::from_handler(handler);
-            self
-        }
-
-        paste! {
-        #[doc = concat!("Route `", stringify!($method) ,"` requests to the given service.")]
-        pub fn [<$method _service>]<S>(mut self, service: S) -> MethodRouter<B, E>
-        where
-            for<'a> S: Service<ServerContext, ServerRequest<B>, Error = E>
-                + Send
-                + Sync
-                + 'a,
-            S::Response: IntoResponse,
-        {
-            self.$method = MethodEndpoint::from_service(service);
-            self
-        }
-        }
-        )+
-    };
-}
-
-impl<B, E> MethodRouter<B, E>
-where
-    B: Send + 'static,
-    E: IntoResponse + 'static,
-{
-    for_all_methods!(impl_method_register_for_builder);
-
-    /// Set a fallback handler for the route.
-    ///
-    /// If there is no method that the route can handle, method router will call the fallback
-    /// handler.
-    ///
-    /// Default is returning "405 Method Not Allowed".
-    pub fn fallback<H, T>(mut self, handler: H) -> Self
-    where
-        for<'a> H: Handler<T, B, E> + Clone + Send + Sync + 'a,
-        T: 'static,
-    {
-        self.fallback = Fallback::from_handler(handler);
-        self
-    }
-
-    /// Set a fallback service for the route.
-    ///
-    /// If there is no method that the route can handle, method router will call the fallback
-    /// service.
-    ///
-    /// Default is returning "405 Method Not Allowed".
-    pub fn fallback_service<S>(mut self, service: S) -> Self
-    where
-        for<'a> S: Service<ServerContext, ServerRequest<B>, Error = E> + Send + Sync + 'a,
-        S::Response: IntoResponse,
-    {
-        self.fallback = Fallback::from_service(service);
-        self
-    }
-}
-
-macro_rules! impl_method_register {
-    ($( $method:ident ),*) => {
-        $(
-        #[doc = concat!("Route `", stringify!($method) ,"` requests to the given handler.")]
-        pub fn $method<H, T, B, E>(handler: H) -> MethodRouter<B, E>
-        where
-            for<'a> H: Handler<T, B, E> + Clone + Send + Sync + 'a,
-            T: 'static,
-            B: Send + 'static,
-            E: IntoResponse + 'static,
-        {
-            MethodRouter {
-                $method: MethodEndpoint::from_handler(handler),
-                ..Default::default()
-            }
-        }
-
-        paste! {
-        #[doc = concat!("Route `", stringify!($method) ,"` requests to the given service.")]
-        pub fn [<$method _service>]<S, B, E>(service: S) -> MethodRouter<B, E>
-        where
-            for<'a> S: Service<ServerContext, ServerRequest<B>, Error = E>
-                + Send
-                + Sync
-                + 'a,
-            S::Response: IntoResponse,
-            B: Send + 'static,
-            E: IntoResponse + 'static,
-        {
-            MethodRouter {
-                $method: MethodEndpoint::from_service(service),
-                ..Default::default()
-            }
-        }
-        }
-        )+
-    };
-}
-
-for_all_methods!(impl_method_register);
-
-/// Route any method to the given handler.
-pub fn any<H, T, B, E>(handler: H) -> MethodRouter<B, E>
-where
-    for<'a> H: Handler<T, B, E> + Clone + Send + Sync + 'a,
-    T: 'static,
-    B: Send + 'static,
-    E: IntoResponse + 'static,
-{
-    MethodRouter {
-        fallback: Fallback::from_handler(handler),
-        ..Default::default()
-    }
-}
-
-/// Route any method to the given service.
-pub fn any_service<S, B, E>(service: S) -> MethodRouter<B, E>
-where
-    for<'a> S: Service<ServerContext, ServerRequest<B>, Error = E> + Send + Sync + 'a,
-    S::Response: IntoResponse,
-    B: Send + 'static,
-    E: IntoResponse + 'static,
-{
-    MethodRouter {
-        fallback: Fallback::from_service(service),
-        ..Default::default()
-    }
-}
-
-#[derive(Default)]
-enum MethodEndpoint<B = Incoming, E = Infallible> {
-    #[default]
-    None,
-    Route(Route<B, E>),
-}
-
-impl<B, E> MethodEndpoint<B, E>
-where
-    B: Send + 'static,
-{
-    fn from_handler<H, T>(handler: H) -> Self
-    where
-        for<'a> H: Handler<T, B, E> + Clone + Send + Sync + 'a,
-        T: 'static,
-        E: 'static,
-    {
-        Self::from_service(handler.into_service())
-    }
-
-    fn from_service<S>(service: S) -> Self
-    where
-        for<'a> S: Service<ServerContext, ServerRequest<B>, Error = E> + Send + Sync + 'a,
-        S::Response: IntoResponse,
-    {
-        Self::Route(Route::new(
-            service.map_response(IntoResponse::into_response),
-        ))
-    }
-
-    fn map<F, B2, E2>(self, f: F) -> MethodEndpoint<B2, E2>
-    where
-        F: FnOnce(Route<B, E>) -> Route<B2, E2> + Clone + 'static,
-    {
-        match self {
-            Self::None => MethodEndpoint::None,
-            Self::Route(route) => MethodEndpoint::Route(f(route)),
-        }
-    }
-}
-
-enum Fallback<B = Incoming, E = Infallible> {
-    Route(Route<B, E>),
-}
-
-impl<B, E> Service<ServerContext, ServerRequest<B>> for Fallback<B, E>
-where
-    B: Send,
-{
-    type Response = ServerResponse;
-    type Error = E;
-
-    async fn call(
-        &self,
-        cx: &mut ServerContext,
-        req: ServerRequest<B>,
-    ) -> Result<Self::Response, Self::Error> {
-        match self {
-            Self::Route(route) => route.call(cx, req).await,
-        }
-    }
-}
-
-impl<B, E> Fallback<B, E>
-where
-    B: Send + 'static,
-    E: 'static,
-{
-    fn from_status_code(status: StatusCode) -> Self {
-        Self::from_service(RouteForStatusCode::new(status))
-    }
-
-    fn from_handler<H, T>(handler: H) -> Self
-    where
-        H: Handler<T, B, E> + Clone + Send + Sync + 'static,
-        T: 'static,
-    {
-        Self::from_service(handler.into_service())
-    }
-
-    fn from_service<S>(service: S) -> Self
-    where
-        S: Service<ServerContext, ServerRequest<B>, Error = E> + Send + Sync + 'static,
-        S::Response: IntoResponse,
-    {
-        Self::Route(Route::new(
-            service.map_response(IntoResponse::into_response),
-        ))
-    }
-
-    fn map<F, B2, E2>(self, f: F) -> Fallback<B2, E2>
-    where
-        F: FnOnce(Route<B, E>) -> Route<B2, E2> + Clone + 'static,
-    {
-        match self {
-            Self::Route(route) => Fallback::Route(f(route)),
-        }
-    }
-
-    fn layer<L, B2, E2>(self, l: L) -> Fallback<B2, E2>
-    where
-        L: Layer<Route<B, E>> + Clone + Send + Sync + 'static,
-        L::Service: Service<ServerContext, ServerRequest<B2>, Error = E2> + Send + Sync + 'static,
-        <L::Service as Service<ServerContext, ServerRequest<B2>>>::Response: IntoResponse,
-        B2: 'static,
-    {
-        self.map(move |route: Route<B, E>| {
-            Route::new(
-                l.clone()
-                    .layer(route)
-                    .map_response(IntoResponse::into_response),
-            )
-        })
-    }
-}
-
-struct StripPrefixLayer;
-
-impl<S> Layer<S> for StripPrefixLayer {
-    type Service = StripPrefix<S>;
-
-    fn layer(self, inner: S) -> Self::Service {
-        StripPrefix { inner }
-    }
-}
-
-struct StripPrefix<S> {
-    inner: S,
-}
-
-impl<S, B, E> Service<ServerContext, ServerRequest<B>> for StripPrefix<S>
-where
-    S: Service<ServerContext, ServerRequest<B>, Response = ServerResponse, Error = E>,
-{
-    type Response = ServerResponse;
-    type Error = E;
-
-    fn call(
-        &self,
-        cx: &mut ServerContext,
-        mut req: ServerRequest<B>,
-    ) -> impl Future<Output = Result<Self::Response, Self::Error>> + Send {
-        let mut uri = String::from("/");
-        if cx
-            .params()
-            .last()
-            .is_some_and(|(k, _)| k == NEST_CATCH_PARAM_NAME)
-        {
-            uri += cx.params_mut().pop().unwrap().1.as_str();
-        };
-        if let Some(query) = req.uri().query() {
-            uri.push('?');
-            uri.push_str(query);
-        }
-
-        // SAFETY: The value is from a valid uri, so it can also be converted into
-        // a valid uri safely.
-        *req.uri_mut() = Uri::from_str(&uri).expect("infallible: stripped uri is invalid");
-        self.inner.call(cx, req)
-    }
-}
-
-struct RouteForStatusCode<B, E> {
-    status: StatusCode,
-    _marker: PhantomData<fn(B, E)>,
-}
-
-impl<B, E> Clone for RouteForStatusCode<B, E> {
-    fn clone(&self) -> Self {
-        Self {
-            status: self.status,
-            _marker: self._marker,
-        }
-    }
-}
-
-impl<B, E> RouteForStatusCode<B, E> {
-    fn new(status: StatusCode) -> Self {
-        Self {
-            status,
-            _marker: PhantomData,
-        }
-    }
-}
-
-impl<B, E> Service<ServerContext, ServerRequest<B>> for RouteForStatusCode<B, E>
-where
-    B: Send,
-{
-    type Response = ServerResponse;
-    type Error = E;
-
-    async fn call(
-        &self,
-        _: &mut ServerContext,
-        _: ServerRequest<B>,
-    ) -> Result<Self::Response, Self::Error> {
-        Ok(self.status.into_response())
-    }
-}
-
 #[cfg(test)]
-mod route_tests {
+mod router_tests {
     use faststr::FastStr;
     use http::{method::Method, status::StatusCode, uri::Uri};
 
-    use super::{any, get, head, options, MethodRouter};
+    use super::Router;
     use crate::{
         body::{Body, BodyConversion},
-        server::{param::PathParamsVec, test_helpers::TestServer},
-        Router, Server,
+        server::{
+            param::PathParamsVec, route::method_router::any, test_helpers::TestServer, Server,
+        },
     };
 
     async fn always_ok() {}
     async fn teapot() -> StatusCode {
         StatusCode::IM_A_TEAPOT
-    }
-
-    #[tokio::test]
-    async fn method_router() {
-        async fn test_all_method<F>(router: MethodRouter<Option<Body>>, filter: F)
-        where
-            F: Fn(Method) -> bool,
-        {
-            let methods = [
-                Method::GET,
-                Method::POST,
-                Method::PUT,
-                Method::DELETE,
-                Method::HEAD,
-                Method::OPTIONS,
-                Method::CONNECT,
-                Method::PATCH,
-                Method::TRACE,
-            ];
-            for m in methods {
-                assert_eq!(
-                    router
-                        .call_route(m.clone(), None)
-                        .await
-                        .status()
-                        .is_success(),
-                    filter(m)
-                );
-            }
-        }
-
-        test_all_method(get(always_ok), |m| m == Method::GET).await;
-        test_all_method(head(always_ok), |m| m == Method::HEAD).await;
-        test_all_method(any(always_ok), |_| true).await;
-    }
-
-    #[tokio::test]
-    async fn method_fallback() {
-        async fn test_all_method<F>(router: MethodRouter<Option<Body>>, filter: F)
-        where
-            F: Fn(Method) -> bool,
-        {
-            let methods = [
-                Method::GET,
-                Method::POST,
-                Method::PUT,
-                Method::DELETE,
-                Method::HEAD,
-                Method::OPTIONS,
-                Method::CONNECT,
-                Method::PATCH,
-                Method::TRACE,
-            ];
-            for m in methods {
-                assert_eq!(
-                    router.call_route(m.clone(), None).await.status() == StatusCode::IM_A_TEAPOT,
-                    filter(m)
-                );
-            }
-        }
-
-        test_all_method(get(always_ok).fallback(teapot), |m| m != Method::GET).await;
-        test_all_method(options(always_ok).fallback(teapot), |m| {
-            m != Method::OPTIONS
-        })
-        .await;
-        test_all_method(any(teapot), |_| true).await;
     }
 
     #[tokio::test]
