+++ volo-http/src/server/utils/mod.rs
@@ -1,11 +1,10 @@
-//! Utilities at server-side
+//! Utilities at server-side.
 
 mod file_response;
 mod serve_dir;
 
 pub use file_response::FileResponse;
 pub use serve_dir::ServeDir;
-
 #[cfg(feature = "ws")]
 pub mod ws;
 #[cfg(feature = "ws")]
