+++ volo-http/src/utils/consts.rs
@@ -1,2 +1,6 @@
+//! Constants of HTTP(S) protocol.
+
+/// Default port of HTTP server.
 pub const HTTP_DEFAULT_PORT: u16 = 80;
+/// Default port of HTTPS server.
 pub const HTTPS_DEFAULT_PORT: u16 = 443;
