+++ volo-http/src/utils/cookie.rs
@@ -1,4 +1,4 @@
-//! Cookie utilities of Volo-HTTP
+//! Cookie utilities of Volo-HTTP.
 //!
 //! [`CookieJar`] currently only supports the server side.
 
@@ -11,7 +11,7 @@ use crate::context::ServerContext;
 #[cfg(feature = "server")]
 use crate::server::extract::FromContext;
 
-/// A cooke jar that can be extracted from a handler
+/// A cooke jar that can be extracted from a handler.
 pub struct CookieJar {
     inner: cookie::CookieJar,
 }
