+++ volo-http/src/utils/mod.rs
@@ -1,3 +1,11 @@
-#![allow(unused)]
+//! Utilities of Volo-HTTP.
+
 pub mod consts;
-pub mod macros;
+#[cfg(feature = "cookie")]
+pub mod cookie;
+mod extension;
+#[cfg(feature = "json")]
+pub(crate) mod json;
+pub(crate) mod macros;
+
+pub use self::extension::Extension;
